+++ spring-core-beans/src/test/java/com/learning/springboot/springcorebeans/part05_aot_and_real_world/serviceloader/DemoGreetingService.java
@@ -1,0 +1,7 @@
+package com.learning.springboot.springcorebeans.part05_aot_and_real_world.serviceloader;
+
+public interface DemoGreetingService {
+
+    String hello();
+}
+