+++ spring-core-beans/src/test/java/com/learning/springboot/springcorebeans/part05_aot_and_real_world/serviceloader/HiGreetingService.java
@@ -1,0 +1,10 @@
+package com.learning.springboot.springcorebeans.part05_aot_and_real_world.serviceloader;
+
+public class HiGreetingService implements DemoGreetingService {
+
+    @Override
+    public String hello() {
+        return "hi";
+    }
+}
+